Move callsite! macro out of core
